+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-05T14:31:06.324Z",
+    "2026-03-06T06:12:43.092Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {