+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3599,7 +3599,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KJZ6J2SPJR976554V1PQKJCY"),
+  id: z.string().default("01KK0WE8B2G33CQBHEEGHE20C0"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),